Change as-list to use namespaced keyword for its type
When writing documentation, it was found that using :type as the option
key was too broad of a name. Changing it to :list/type makes it very
specific what is is about, and makes it possible to use it in the
default options, without fear of it impeding something else that might
need to use a :type keyword.