Update flightform example with i18n and better validation
The labels don't update on the fly with the change in re-frame. Should
be looked into if labels, text, etc can be done better. The i18n
protocol is extended to make use of tongue as the i18n library. A
ranged datepicker from antd is used and the validation demonstrates
the use of function based validations.