Add reset-form! to helpers
Mirrored reset-form! to ez-wire.form.

The implementation turned out to be a bit tricky, and the order in
which things are done matter quite a bit. Clearing the errors are
needed, so that we can always start from a clean slate.

For certain components using :default-value instead of :value is
crucial, since there are problems with cursor placement when using
:value for such elements. Using :default-value eliminates that
particular problem, but in turn a reset of the :data RAtom in the
field does nothing.

A new key named form-key has been added to solve this
problem. form-key is a RAtom that changes when reset-form! is used,
thus forcing a complete re-render of the form, which forces the
re-render of any field that happens to use :default-value instead of :value.