Add new module ui to ez-wire
There is a need for a seperate UI module in ez-wire for admin. While
admin should be completely replacable for components if the user so
wishes to do so, there is no need to burden ez-wire with a hard
dependancy on other component libraries.

At the same time it would be highly helpful to have a handful of
components that can be used in the admin interface as a default
implementation, that is not hardwired to anything outside of ez-wire.

It is typically hard to write good, generic, UI components. One day
the UI module might be worthwhile, but for the time being it is better
to keep it small.