Add optional override for automatic cleanup for forms
There is an automatic cleanup of forms in all the rendering of the
forms. This works well up until the point where a re-render of the
reagent component where the form is rendered happens. At that point
the data is cleaned up in re-frame, and the form is now invalid.

To combat this, it's now possible to turn off the automatic cleanup,
and let it be handled manually.