Add sorting to table
Can now sort reliably between different columns. Reshuffled a bit in
the setup of things, and is now moving the context around. :row/
namespace is a bit confusing, as it is used for what really are cells,
as opposed to the row specifically. key-fn for rows needs to be added,
and should use an index as default, but be overridable.