Add multi-validation to validation
Allows for a validation for one field using multiple fields in the
form to perform the validation. Had to add a new function to the
IValidate protocol called get-affected-fields, so that the
get-validation-errors function in form could check if it needs to
update, even when other affected fields are updated, rather than the
field the validation is attached to.