Add pagination and default index to table
Index for the table row uses :row/key-fn. In the event that does not
exist, init-context adds an index to each row, useable by the functions
responsible for handling the rendering.

Pagination has also been added to the table. As of yet no pagination
element has been added to ui, so that is still missing.