Update form for branching (plus MultiValidation fix)
MultiValidation was slightly broken. Changing the check to just
checking that the current field has some kind of value, and then
running field-updated? (previously called should-update-error?) on all
the fields affected fields (plus current) seems to do the trick. There
was an edge case, where an update to an affected field wouldn't
trigger the check for the affecting field.

Branching has been added. It's working, but looks to be easily abused
for a complete maze of updates. The implementation is very free with
what you can do, and enforces very little.