Add elements and update test page
Added branching to the test page. :value vs :default-value is a
problem for text fields. The form could avoid most of the problems
when adapting React components that use text fields, by simply using
:default-value. With the branching, however, it breaks. When not
showing a field in the form, React completely removes it from the DOM,
causing the any re-emergance of the field to display the
:default-value, despite the :value having been updated before.

Cannot find a way around it, other than letting the component in the
field correctly handling setting :value.