Add branching initiation check for the form watcher
When we initiate a form, we run the watcher manually by first having
the :data RAtom set to nil for all values, and then reset the data
with the input data. This triggered the branching as well, whenever
the data in fields that was covered by branching was not nil. A small
check in the watcher handles this, and any subsequent updates to the
form runs against the branching logic as normal.